Cherry-pick improvements from dev branch

+++ src/DocSharp.Docx/Helpers/ListHelpers.cs
@@ -28,7 +28,7 @@ public static class ListHelpers
         return uppercase ? result.ToUpperInvariant() : result;
     }
 
-    public static string NumberToRomanLetter(int number, bool uppercase)
+    public static string NumberToRomanLetter(long number, bool uppercase)
     {
         if (number < 1)
             number = 1;
